+++ src/Dfe.PlanTech.Web/Views/Shared/Questionnaire/Category.cshtml
@@ -19,7 +19,7 @@
 
         <task-list-item>
             <task-list-item-name>
-                <a href=@categorySection.Name class="govuk-link">
+                <a asp-controller="Pages" asp-action="Get" class="govuk-link">
                     @categorySection.Name
                 </a>
             </task-list-item-name>
